Fix file restore from backward diffs
